refactor(utils): move date-related code to utils.datetime
