+++ ggshield/core/text_utils.py
@@ -1,5 +1,4 @@
 import sys
-from datetime import datetime
 from typing import Any, Dict, Optional, Union
 
 import click
@@ -193,11 +192,3 @@ def file_diff_info(
         format_text(filename, STYLE["filename"]),
         ", ".join(incidents_count),
     )
-
-
-def get_pretty_date(dt: datetime) -> str:
-    """
-    convert the given datetime to the format September 1, 2022
-    """
-    # Don't use %d for the day because it adds a leading 0
-    return dt.strftime(f"%B {dt.day}, %Y")
